+++ deeplens/optics/geometric_surface/base.py
@@ -7,7 +7,7 @@ import numpy as np
 import torch
 import torch.nn.functional as F
 
-from deeplens.optics import DeepObj
+from deeplens.optics.base import DeepObj
 from deeplens.optics.config import EPSILON
 from deeplens.optics.material import Material
