+++ deeplens/optics/phase_surface/phase.py
@@ -6,7 +6,7 @@ import torch
 import torch.nn.functional as F
 
 from deeplens.optics.config import EPSILON
-from deeplens.optics import DeepObj
+from deeplens.optics.base import DeepObj
 from deeplens.optics.material import Material
